website-django/requirements.txt
asgiref==3.5.0
Django==4.0.4
djangorestframework==3.13.1
docutils==0.18.1
Pillow==9.1.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
